requirements.txt
aiohttp
python-chess
pandas